Fix: 3가지를 수정함. 아래 커밋메세지에서 확인 가능
server.cpp 파일을 각각 용도에 따라서 나눔
---
Fix: configfile에서 autoindex true/false -> on/off로 변경
---
Feat: configfile에 method, cgi_extension 추가

method를 추가하여, OPTIOINS 에서 사용할 수 있을 것 같고,
cgi_extension을 추가하여, 해당 폴더에서 어떤 확장자의 파일을 지원하는지 설정해보았다.

make re를 해서,  this->ShowServers() 함수로 확인해보면 좋음
