+++ default.config
@@ -4,40 +4,19 @@ server
 	server_name localhost
 	error_page error.html
 	index_pages index.html index2.html
+	index_pages sdss.html
 	root /
-	auto_index true
+	auto_index on
+	method GET POST PUT DELETE
+	cgi_extension .bla
 	location /abc
 	{
 		root ef
 		index_pages index_abc.html
 		error_page error_abc.html
+		method GET
+		method TTT
+		cgi_extension .bla .php
+		cgi_extension .cpp .php
 	}
 }
-
-server
-{
-	port 8001
-	server_name hello
-	error_page error2.html
-	index_pages index.html
-	auto_index true
-	timeout 200
-}
-
-server
-{
-	port 8002
-	error_page error2.html
-	index_pages index.html
-}
-
-server
-{
-	port 8002
-	server_name yunslee
-	error_page error.html
-	index_pages index.html
-	auto_index true
-	timeout 10
-}
-
