+++ default2.config
@@ -1,0 +1,43 @@
+server
+{
+	port 8000
+	server_name localhost
+	error_page error.html
+	index_pages index.html index2.html
+	root /
+	auto_index on
+	location /abc
+	{
+		root ef
+		index_pages index_abc.html
+		error_page error_abc.html
+	}
+}
+
+server
+{
+	port 8001
+	server_name hello
+	error_page error2.html
+	index_pages index.html
+	auto_index on
+	timeout 200
+}
+
+server
+{
+	port 8002
+	error_page error2.html
+	index_pages index.html
+}
+
+server
+{
+	port 8002
+	server_name yunslee
+	error_page error.html
+	index_pages index.html
+	auto_index on
+	timeout 10
+}
+
