+++ srcs/ConfigFile/ConfigFiles.cpp
@@ -19,6 +19,18 @@ void ConfigFiles::ShowConfigs()
 		}
 		std::cout << std::endl;
 		std::cout << "error_page: " << this->mconfigs[i].merror_page.getPath() << std::endl;
+		std::cout << "method: ";
+		for (size_t j = 0; j < this->mconfigs[i].m_method.size(); j++)
+		{
+			std::cout << this->mconfigs[i].m_method[j] << " ";
+		}
+		std::cout << std::endl;
+		std::cout << "cgi_extension: ";
+		for (size_t j = 0; j < this->mconfigs[i].mcgi_extension.size(); j++)
+		{
+			std::cout << this->mconfigs[i].mcgi_extension[j] << " ";
+		}
+		std::cout << std::endl;
 	}
 	std::cout << "=========================================================" << std::endl;
 }
