+++ srcs/Server_run.cpp
@@ -1,0 +1,231 @@
+#include "Server.hpp"
+#include "ServerManager.hpp" // NOTE 상호참조 문제를 해결하기 위해서!
+#include "Response.hpp"
+
+extern char**	g_env;
+
+void		Server::run(void)
+{
+	std::map<int, Connection>::iterator	it = m_connections.begin();
+	while (it != m_connections.end())
+	{
+		std::map<int, Connection>::iterator it2 = it++;
+		int fd = this->msocket;
+		if (it2->second.get_m_fd() == fd)
+		{
+			continue ;
+		}
+		try
+		{
+			if (hasSendWork(it2->second))
+			{
+			 	runSend(it2->second);
+			 	continue ; // FIXME 어떻게 처리할지...
+			}
+			if (hasExecuteWork(it2->second))
+			{
+				runExecute(it2->second);
+				continue ;
+			}
+			if (hasRequest(it2->second))
+			{
+				if (it2->second.get_m_request() == NULL)
+				{
+					it2->second.set_m_request(new Request());
+				}
+				runRecvAndSolve(it2->second);
+			}
+		}
+		// catch (Server::IOError& e)
+		// {
+		// 	ft::log(ServerManager::log_fd, ft::getTimestamp() + e.location() + std::string("\n"));
+		// 	closeConnection(fd);
+		// }
+		catch (const std::exception& e)
+		{
+			cout << e.what() << endl;
+			// ft::log(ServerManager::log_fd, ft::getTimestamp() + "detected some error" + std::string("\n"));
+			closeConnection(it2->second.get_m_fd());
+
+		}
+	}
+	if (hasNewConnection())
+	{
+		if (m_connections.size() >= (INIT_FD_MAX / m_manager->GetServers().size()))
+		{
+			int fd = getUnuseConnectionFd();
+			if (fd == -1)
+				return ;
+			closeConnection(fd);
+		}
+		
+		if (false == acceptNewConnection())
+		{
+			cout << "accpetNewconnection(): Fail" << endl;
+			// reportCreateNewConnectionLog();
+		}
+		else
+		{
+			cout << "accpetNewconnection(): Success" << endl;
+		}
+	}
+	// cout << this->mport << "'s connection_size: "<< m_connections.size() << endl; 
+}
+
+bool		Server::hasSendWork(Connection& connection)
+{
+	// value = connection.get_m_request()->GetPhase();
+	// TODO COMPLETE로 가정하였으나, 실제로는 request의 진행상황에 따라서 { READY, ON_HEADER, ON_BODY, COMPLETE }; 단계로 나뉨.
+	if (connection.get_m_request() == NULL)
+		return (false);
+	Request::ePhase value;
+	value = connection.get_m_request()->GetPhase();
+	
+	if (value == connection.get_m_request()->COMPLETE)
+	{
+		if (FD_ISSET(connection.get_m_fd(), &(this->m_manager->GetWriteCopySet())) <= 0)
+		{
+			closeConnection(connection.get_m_fd());
+			return (false);
+		}
+		return (true);
+	}
+	else
+		return (false);
+}
+
+bool		Server::runSend(Connection& connection)
+{
+	int clinet_socket = connection.get_m_fd();
+	Request *request = connection.get_m_request();
+	char buffer[100];
+
+	bool send_complete = false;
+	if (connection.GetStatus() == Connection::SEND_READY)
+	{
+		write(connection.get_m_fd(), connection.get_m_response()->getResponse().c_str(), connection.get_m_response()->getResponse().size());
+		// cout << connection.get_m_response()->getResponse() << endl;
+		send_complete = true;
+	}
+	closeConnection(connection.get_m_fd());
+	return (send_complete);
+}
+
+bool		Server::hasExecuteWork(const Connection& connection) const
+{
+	if (connection.GetStatus() == Connection::CGI_READY ||
+		connection.GetStatus() == Connection::SEND_READY || // STUB	나중에 지워야함
+		connection.GetStatus() == Connection::CGI_ING)
+	{
+		return (true);
+	}
+	else
+	{
+		return (false);
+	}
+}
+
+bool		Server::runExecute(const Connection& connection)
+{
+	if (connection.GetStatus() == Connection::CGI_READY)
+	{
+		createCGIEnv(connection);
+	}
+
+	// executeCGI();
+	return (false);
+}
+
+bool		Server::hasRequest(const Connection& connection)
+{
+	if (FD_ISSET(connection.get_m_fd(), &(this->m_manager->GetReadCopySet()))) // REVIEW	request의 phase도 함께 확인해야할 수도 있을 듯
+	{
+		// std::cout << "client(" << connection.get_m_fd() << ") : has request" << std::endl;
+		return (true);
+	}
+	else
+	{
+		return (false);
+	}
+}
+
+bool		Server::runRecvAndSolve(Connection& connection)
+{
+	try
+	{
+		recvRequest(connection);
+	}
+	catch (int status_code)
+	{
+		std::cout << "status code : " << status_code << std::endl;
+		create_statuspage_Response(connection, status_code);
+		connection.SetStatus(Connection::SEND_READY);
+		return (true);
+	}
+	// catch (const Server::IOError& e)
+	// {
+	// 	throw (e);
+	// }
+	catch (const std::exception& e)
+	{
+		// ft::log(ServerManager::log_fd, std::string("[Failed][Request] Failed to create request because ") + e.what());
+		// createResponse(connection, 50001);
+	}
+
+	try
+	{
+		Request& request = *connection.get_m_request();
+		if (request.GetPhase() == Request::COMPLETE)
+		{
+		// 	writeCreateNewRequestLog(request);
+		// 	connection.set_m_status(Connection::ON_EXECUTE); //REVIEW 이게 맞나?
+			solveRequest(connection, *connection.get_m_request());
+			return (true);
+		}
+		return (false);
+	}
+	catch (int status_code)
+	{
+		std::cout << "status code : " << status_code << std::endl;
+		create_statuspage_Response(connection, status_code);
+		connection.SetStatus(Connection::SEND_READY);
+	}
+	catch(const std::exception& e)
+	{
+		std::cerr << e.what() << '\n';
+	}
+	return (false);
+}
+
+bool		Server::hasNewConnection()
+{
+	if (FD_ISSET(this->msocket, &(this->m_manager->GetReadCopySet())))
+	{
+		cout << "this->msocket: " << this->msocket << endl;
+		return (true);
+	}
+	else
+	{
+		return (false);
+	}
+}
+
+bool		Server::acceptNewConnection()
+{
+	int client_socket;
+	sockaddr_in sockaddr;
+	socklen_t socketlen;
+	socketlen = sizeof(struct sockaddr);
+	client_socket = accept(this->msocket, (struct sockaddr*)&sockaddr, (socklen_t*)&socketlen);
+	if (client_socket == -1)
+	{
+		std::cerr << "Could not create socket." << std::endl;
+		return (false);
+	}
+	fcntl(client_socket, F_SETFL, O_NONBLOCK);
+	FD_SET(client_socket, &(this->m_manager->GetReadSet()));
+	FD_SET(client_socket, &(this->m_manager->GetWriteSet()));
+	this->m_connections[client_socket] = Connection(client_socket, "", this->mport);
+	// close(client_socket); // NOTE 이제 keep-alive로 관리
+	return (true);
+}
